Return correct uri for TypeMergeError exceptions with dictionaries
